Remove deprecation notice
After talking with @trotttrotttrott we came to the conclusion that the
next generation API client won't be available any time soon, and we
continue to accept contributions to it from time to time.

Note that removing the deprecation warning doesn't mean we have a
roadmap for upgrades, this just removes the deprecation notice that
might scare the community from sending contributions.

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>